refactor: move auth APIs to a sedicated service handler
